fix(apix): generation time ms is a float
Caused issues in the protobuf de-serialization

+++ lib/src/main/kotlin/com/openmeteo/apix/common/response/ResponseGenerationTimed.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.SerialName
 
 interface ResponseGenerationTimed : Response {
     @SerialName("generationtime_ms")
-    val generationTimeMs: Double
+    val generationTimeMs: Float
 }
